Remove extra xmrig exit listener

+++ src/index.ts
@@ -593,17 +593,8 @@ io.on('connection', (socket) => {
   socket.on('start_mining', (config) => {
     log.info('socket event: start_mining');
     log.info('Starting mining process');
-
-    if (xmrigProcess) {
-      xmrigProcess.kill();
-      xmrigProcess.on('exit', () => {
-        log.info('Previous mining process terminated.');
-        xmrigProcess = null;
-        startMining(config);
-      });
-    } else {
-      startMining(config);
-    }
+    if (xmrigProcess) xmrigProcess.kill();
+    startMining(config);
   });
   socket.on('stop_mining', () => {
     log.info('socket event: stop_mining');
